refactor tests common setup
