+++ .secrets.baseline
@@ -248,7 +248,7 @@
         "filename": "tests/integration/datadict/submission/test_endpoints.py",
         "hashed_secret": "78b4db9b2aec0f0f2d3e38f9278be42b861c9dc3",
         "is_verified": false,
-        "line_number": 1457
+        "line_number": 1396
       }
     ],
     "tests/integration/datadictwithobjid/submission/data/biospec1.json": [
@@ -327,5 +327,5 @@
       }
     ]
   },
-  "generated_at": "2023-01-25T23:25:56Z"
+  "generated_at": "2023-01-26T23:32:07Z"
 }
