+++ tests/integration/datadict/conftest.py
@@ -1,3 +1,4 @@
+import importlib
 import os
 import json
 import multiprocessing
@@ -19,17 +20,15 @@ from sheepdog.test_settings import (
 )
 
 from tests import utils
-from tests.integration.datadict.api import app as _app, app_init, indexd_init
-from tests.integration.datadict.submission.test_endpoints import put_cgci_blgsp
-import importlib
+from tests.integration.utils import (
+    get_parent,
+    wait_for_indexd_alive,
+    wait_for_indexd_not_alive,
+)
+from tests.integration.api import app as _app, app_init, indexd_init
 
 
 multiprocessing.set_start_method("fork")
-
-
-def get_parent(path):
-    print(path)
-    return path[0 : path.rfind("/")]
 
 
 PATH_TO_SCHEMA_DIR = (
@@ -70,26 +69,6 @@ def require_index_exists_on(app, monkeypatch):
 @pytest.fixture
 def require_index_exists_off(app, monkeypatch):
     monkeypatch.setitem(app.config, "REQUIRE_FILE_INDEX_EXISTS", False)
-
-
-def wait_for_indexd_alive(port):
-    url = "http://localhost:{}/_status".format(port)
-    try:
-        requests.get(url)
-    except requests.ConnectionError:
-        return wait_for_indexd_alive(port)
-    else:
-        return
-
-
-def wait_for_indexd_not_alive(port):
-    url = "http://localhost:{}/_status".format(port)
-    try:
-        requests.get(url)
-    except requests.ConnectionError:
-        return
-    else:
-        return wait_for_indexd_not_alive(port)
 
 
 @pytest.fixture
@@ -154,24 +133,6 @@ def app(tmpdir, request):
     return _app
 
 
-@pytest.fixture()
-def use_ssl(request):
-    try:
-        # one of [False, True, None]
-        return request.param
-    except Exception:
-        return None
-
-
-@pytest.fixture()
-def isolation_level(request):
-    try:
-        # one of ["READ_COMMITTED", "REPEATABLE_READ", "SERIALIZABLE", None]
-        return request.param
-    except Exception:
-        return None
-
-
 @pytest.fixture
 def pg_driver(request, client, use_ssl, isolation_level):
     pg_driver = PsqlGraphDriver(
@@ -196,11 +157,6 @@ def pg_driver(request, client, use_ssl, isolation_level):
     tearDown()
     request.addfinalizer(tearDown)
     return pg_driver
-
-
-@pytest.fixture()
-def cgci_blgsp(client, submitter):
-    put_cgci_blgsp(client, submitter)
 
 
 @pytest.fixture()
