+++ tests/integration/datadictwithobjid/conftest.py
@@ -16,17 +16,16 @@ from gen3authz.client.arborist.client import ArboristClient
 
 
 from sheepdog.test_settings import INDEX_CLIENT
-from tests.integration.datadictwithobjid.api import app as _app, app_init, indexd_init
-from tests.integration.datadictwithobjid.submission.test_endpoints import put_cgci_blgsp
+from tests.integration.utils import (
+    get_parent,
+    wait_for_indexd_alive,
+    wait_for_indexd_not_alive,
+)
+from tests.integration.api import app as _app, app_init, indexd_init
 from tests import utils
 
 
 multiprocessing.set_start_method("fork")
-
-
-def get_parent(path):
-    print(path)
-    return path[0 : path.rfind("/")]
 
 
 PATH_TO_SCHEMA_DIR = (
@@ -58,36 +57,6 @@ def pg_config(use_ssl=False, isolation_level=None):
         ret_val["isolation_level"] = isolation_level
 
     return ret_val
-
-
-@pytest.fixture
-def require_index_exists_on(app, monkeypatch):
-    monkeypatch.setitem(app.config, "REQUIRE_FILE_INDEX_EXISTS", True)
-
-
-@pytest.fixture
-def require_index_exists_off(app, monkeypatch):
-    monkeypatch.setitem(app.config, "REQUIRE_FILE_INDEX_EXISTS", False)
-
-
-def wait_for_indexd_alive(port):
-    url = "http://localhost:{}/_status".format(port)
-    try:
-        requests.get(url)
-    except requests.ConnectionError:
-        return wait_for_indexd_alive(port)
-    else:
-        return
-
-
-def wait_for_indexd_not_alive(port):
-    url = "http://localhost:{}/_status".format(port)
-    try:
-        requests.get(url)
-    except requests.ConnectionError:
-        return
-    else:
-        return wait_for_indexd_not_alive(port)
 
 
 @pytest.fixture
@@ -152,24 +121,6 @@ def app(tmpdir, request):
     return _app
 
 
-@pytest.fixture()
-def use_ssl(request):
-    try:
-        # one of [False, True, None]
-        return request.param
-    except Exception:
-        return None
-
-
-@pytest.fixture()
-def isolation_level(request):
-    try:
-        # one of ["READ_COMMITTED", "REPEATABLE_READ", "SERIALIZABLE", None]
-        return request.param
-    except Exception:
-        return None
-
-
 @pytest.fixture
 def pg_driver(request, client, use_ssl, isolation_level):
     pg_driver = PsqlGraphDriver(
@@ -194,11 +145,6 @@ def pg_driver(request, client, use_ssl, isolation_level):
     tearDown()
     request.addfinalizer(tearDown)
     return pg_driver
-
-
-@pytest.fixture()
-def cgci_blgsp(client, submitter):
-    put_cgci_blgsp(client, submitter)
 
 
 @pytest.fixture()
